[icon] a11y (#390)
<!--- Commit was signed off by phytonmk@gmail.com with GPG key ID
CF2DC815AE956C1F -->

Made aria-label required for interactive icons

## Description

If `interactive` prop is provided for icon but `aria-label` or
`aria-labelledby` are not, a warning is logged in non-production env.

Non-interactive icons are aria-hidden now.

## How has this been tested?

I have not added any new tests but updated lot of them (by adding
aria-label). Suppose if now axe errors are appeared and no new warnings
during jest run changes may be considered as working.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.