[core] Children Index definitions (#402)
## Description

I was able to add a cache to the index, which allowed me to get the
correct value when re-rendering. But I had to add another class, since
it is the only one that can store the value of the instance.

I think this is a temporary solution to quickly get used to the new
react. In the future it will be necessary to do this more elegantly.

## Motivation and Context

Need to upgrade to react 18.

## How has this been tested?

Open website and see examples and run tests; Also, update React to 18
version and test website.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.