[neighbor-location] Major update 😬 (#394)
## Description

The approach to determining neighbors has been changed. Now the
neighbors are determined by enumeration of children
  and not by context.
Added a new option for using custom components for users. An example is
on the site.
Updated all dependent components. They didn't get a major.

## Motivation and Context

Need to upgrade to react 18. And there the definition through the
context will not work correctly.

## How has this been tested?

Open website and see examples; Also, update React to 18 version and test
website.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.