[ellipsis] Ellipsis (#398)
Create new component - Ellipsis

## Description

We got a lot of requests to create a new component that will wrap the
text in an ellipsis with a tooltip. It has many use cases and will be
useful to users

## Motivation and Context

Despite the fact that our design system already has many components,
there are still many use cases where they are not enough and developers
need new components to simplify development. We received a large number
of requests when users might need this component - use in table headers,
links, large amounts of text in cards, and so on. Therefore it was
decided to create a new component

## How has this been tested?

Component has been visually tested to work correctly for a variety of
situations. Also, unit tests were written, which check the main use
cases of the component (basic use of the component, the case of
multiline text and trimming in the middle of the text)

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.

Signed-off-by: elena.krasnopolskaia <elena.krasnopolskaia@semrush.com>
Co-authored-by: Julia Mnizhek <j.mnizhek@semrush.com>