[data-table] Table in table with accordion and fixed column (#415)
## Description

There are two problems:
1. When using a table within a table, the calculation of column widths
was out of sync. Therefore, I added one more argument from which you can
take the size of the column.
2. When using a fixed column, the sticks do not work if one of the
parents has an `overlay != visible`. I've added a few properties to
block the overlay from being set by default. The accordion had it for
animation. The table had it for scroll.

## Motivation and Context

New tables don't work with complex use cases. It blocks the transition
to them.

## How has this been tested?

I added a few examples in documentation.
Open the site and show examples. For this, I deployed RC
https://intergalactic-mj3h2dvptq-uk.a.run.app/
You can play in the playground 🧑‍💻

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.

Signed-off-by: Roman <r.lysov@semrush.com>