[ellipsis] supported passing tooltip props (#542)
## Description

Now a defined list of props passed to ellipsis components are being
passed to tooltip. A special prop `includeTooltipProps` allowes user to
redefine a list of passed props.

## Motivation and Context

No ability to pass tooltip props seems weird and unexpected. I've made
passing of props similar to feedback form component and redefining list
of passed props as in checkbox component.

## How has this been tested?

Manual testing was enough. The change is pretty easy.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.