[website] use amplitude device id as user token for algolia (#562)
Added synonyms to site search in Algolia.

## Description

To have possibility to use synonyms in algolia we need the `userToken`
parameter but we don't have users. For amplitude analytics we have the
`deviceId` parameter (a randomly generated string that is stored in
cookies). What happens if we use the `deviceId` parameter as the
`userToken` parameter?