[select] autoscroll fix (#565)
## Description

Fixed automatic scrolling to selected option on popper open wasn't
working and added prop to control should be popper view scrolled to
selected option on it's open.

## Motivation and Context

Previously we had a mechanism that was controlling autoscroll to first
selected option in `<Scroll />` component. Currently it doesn't work. I
have rebuilt mechanism to make it more reliable and also added prop do
disable mechanism as some of our users were requesting.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.

---------

Signed-off-by: Roman <r.lysov@semrush.com>
Co-authored-by: Roman <r.lysov@semrush.com>