[core] Changed logic of merging component styles and context styles due to losing context styles before (#579)
## Description

We encountered a styles issue when using a `DropdownMenu` inside a
`DataTable`. The styles applied to the `ScrollArea` inside the
`DropdownMenu` were not being displayed properly on the `DropdownMenu`
inside the `DataTable`. This occurred because the styles of the
`ScrollArea` had already been overwritten in the `DataTable`. The root
cause of the issue was a bug in the `@semcore/core` package, which
failed to consider the possibility of different styles being present in
the context and the component itself.