[data-table,table] Change layout sort icon (#604)
## Description

By default, sorting runs over the text in any of its directions and at
any column size. The place for the icon is not reserved, so the text is
not pushed anywhere.
I also added line-height for table head as it was in the design. Because
of this, a lot of screenshots have changed.

## Motivation and Context

To reduce the column size

## How has this been tested?

View screenshot tests

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.

---------

Signed-off-by: r.lysov <r.lysov@semrush.com>