[typography] added new property for Text component (#627)
Added new property **uppercase**, **lowercase** and **capitalize** for
Text component

## Description

This is my **proposal** for adding new properties to the Text component.

## Motivation and Context

Very often when developing, you need to do a basic text transformation
and to do this you need to either add a **raw style** or write a
**className**, it seems to me that such basic work with styles should be
available out of the box.

For example like this:
`<Text uppercase>text</Text>`

Perhaps need to use instead of these properties to create just one
(**textTransform** or just **transform**) since this property can only
be one at the same time, like your **decoration**
`<Text transform={'uppercase'|'lowercase'|'capitalize'}>text</Text>`

## How has this been tested?

Just added an example to the documentation and it works :)
Yes, and the tests have fallen for some reason, but there is no rerun
button, but you can just discuss

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.

# If appropriate sections above are not fulfilled properly do not even
review my pull request and never approve it! Only after fulfilling
sections above I will also remove this note.