[popper] fixed keyboard focused poppers were not intractable with mouse (#658)
## Motivation and Context

We have bug – when popper is keyboard focus (e.g. with `Tab`) it becomes
non intractable with mouse. It appeared after
https://github.com/semrush/intergalactic/pull/630 when Popper got
`:after` pseudoelement for focused state highlighting. What caused the
issue is that we forgot to add `pointer-events: none` for that
pseudoelement.

Also I found that TimePicker with `sv` locale wasn't published properly.
Including changelog diff in that PR.

## How has this been tested?

I have added browser tests on that behavior.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.