[input] input now catches focus after keyboard events happened inside… (#659)
… of the input container if no other element on page has focus

## Motivation and Context

There is popular case – Input contains clear button implemented via
`Input.Addon`.

If user clicks this clear button it automatically sets focus on input. 

If user triggers clear button with keyboard, the button disappears from
page and goes to the `docuemnt.body`. I have added keyboard handler on
the input container that checks where the focus is and if focus is on
the `document.body`, it returns focus to the input.


## How has this been tested?

I have added browser tests.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue).
- [x] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.