[global] supported `tag={React.Fragment}` (#690)
## Motivation and Context

Sometimes it might very helpful to pass React.Fragment as a component
tag. Sure, it will disable most of components functionality but it seems
to a an important escape hatch. Currently it works but with React
warnings (e.g. "Invalid prop `className` supplied to `React.Fragment`").

## How has this been tested?

All existing tests should pass well.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.

## Checklist:

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.