[global] switched repo from abandoned rome to it's successor biome (#770)
## Motivation and Context

Several monthes ago I switched repo from prettier and eslint to rome.
And not so long time ago it's got abandoned.

I am switching it to it's successor biome – it has same api and all
migration took like a 10 minutes.

Why I switched to biome and not back to eslint+prettier:
1. I am still not very satisfied with performance of eslint and
prettier.
2. Biome is under the active development – it's even getting JetBrains
IDE support very soon.
3. Migration to biome is really fast because api is the same.  

So, from reasons above I want to give that tool a second chance. I it
fill fuck up as the rome – migration to eslint and prettier would be
inevitable.

## How has this been tested?

No tests :)

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [ ] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.