[utils] React invalid Hooks call caused by `useFocusLock` in some rare cases from it's nested hooks (#781)
## Motivation and Context

`useFocusLock` hook is merging if multiple versions of it is loaded on
one page.
It's critical to provide smooth UX and prevent focus lock wars.

To make it work on page where multiple projects use multiple versions of
React, the React isolation was added. However, it was added with a
mistake – nested hooks wasn't isolating React.

This PR adds such isolation.

## How has this been tested?

It was tested manually and I am going to test prerelease version on real
world project before publishing release.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.