[base-trigger] empty FilterTrigger now gets role `button` during empty state instead of `group` (#787)
## Motivation and Context

Initially filter trigger had `button` role. Then we found that it's
wrong when container has role `button` while having buttons (clear
buttons) inside. So we changed it to `group`.

But action it's not so good. We always needed `button` role when filter
empty has empty state and `group` when it has clear button inside. It
wasn't implemented yet because props overriding hack (prefixing with
`use:`) was working only with style props. Now I made it to work with
any prop and made FilterTrigger role group work as needed.

## How has this been tested?

All tests are passing. Manual tests confirmed that FilterTrigger role is
changing

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.