Small core fixes (#798)
## Motivation and Context

I found several issues.
1. BaseTrigger generates warnings about using non existing attribute.
That would be better fixed after
https://github.com/semrush/intergalactic/issues/748 but for now I just
used the attribute name that is not passing _the grand regexp_.
2. If function is provided to ref prop, it will be called twice in deep
nested components because assignProps were not clearing forwardRef prop
after merging it into ref prop.
3. ScrollArea component remounts after first render. It caused by
assignProp function that generates new ref function callback on every
render. It cannot be fixed with current core and is not fixed in this
PR.

## How has this been tested?

All existing tests should pass.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.