[utils] fixed focus lock merging was causing invalid hooks order error (#839)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

When updating focus lock from version 2 to version 3 `useUniqueId` hook
was removed.
Due to replacing useFocusLock hook in runtime, the order of hooks must
be preserved.
So, useUniqueIdHookMock helps to preserve hooks order when hook is being
replaced from version 2 to version 3.

## How has this been tested?

It was tested on stage that uses multiple versions of `@semcore/ui` and
where both focus lock v2 and focus lock v3 are present.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.