Resolvable design tokens (#811)
https://github.com/semrush/intergalactic/issues/747

The `useResolveColor` hook was added to utils as a replacement to
`resolveColor` function. Instead of converting color key to it's hex
value, it acts differently:
1. If the passed color is taken from the old palette – it sends (only in
dev mode) warning into console and returns color as `resolveColor` do.
2. If passed color is a design token (`--intergalctic-` prefix is
optional). It takes it's default value from default theme file.
3. After getting default value, hook returns css variable expression
with default value that may be used in css.

I made resolving to css variables because it allows absolutely smooth
theme switching as it must work.

With this hook it's not not possible to apply color filters from js as
we used to do in `semcore/utils/src/color.ts` file. So I changed
component that was relying on that mechanism to use only css filters and
properties to change the colors. For example, in `Tag` component the
text color is also applied to background but the background also gets
`opacity: 0.1; filter: brightness(1.5) saturate(1.5);` to make the text
look nice on it. Same applied to badge and counter.

Also, that changes made it possible to use
`--intergalactic-chart-palette-order-{n}` tokens in charts. So I removed
most uses of explicit `color` prop in charts documentation.

I had a lot of troubles with making automatic colors work in the charts'
tooltip because due to some very weird hacks in the `semcore/core` the
`getDotProps` on tooltip wasn't working. I have added additional context
that seems to work nice but absolutely ugly in general.

The PR is ready for review.

---------

Co-authored-by: Julia Mnizhek <j.mnizhek@semrush.com>