[skeleton] fixed missing animation (#847)
## Motivation and Context

It was found that skeleton animation wasn't working, not sure how this
happend in this overcomplecated component, but explicit passing
animation duration in `SkeletonSVG` solves the issue.

## How has this been tested?

It started working as expected in my local playground. I am not sure we
need any kind of video tests.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.