[utils, website] fixed old palette resolving and also fixed palette view on website (#873)
## Motivation and Context

The color resolver contained a very stupid mistake that was making using
old color palette impossible. It has been existing in our codebase for
about two weeks and it's insane that we haven't spotted it.

Basically it was removing `--` prefix from variable name and then was
checking does this variable exist in map where all names had prefix
`--`. So our potential users had to use `----{variable_name}` syntax if
they were trying to overcome the issue.

## How has this been tested?

The fix is clear.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.