Screenshoting in docker (#934)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

Our remote screenshot service is very unstable. It was very aweful DX to
restart screenshot tests like 3 or 5 times per PR or release. I've moved
screenshot testing into docker container.

It may be slower a little bit but it seems to work very-very stable.

## How has this been tested?

With existing tests :)

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.