[input-mask] fixed that component was breaking a page when initial value was not allowed by outer pipe (#980)
## Motivation and Context

It was found that if initial value is not allowed by outer pipe
function, component blows up due to error inside of `input-mask-core`.
Input mask core have no updated last 5 years and migrating to some other
library is very painful (but needed in the future).

Here is an example when it was initially found – selected day is
disabled
https://codesandbox.io/p/sandbox/frosty-cloud-25yvw7?file=%2Fsrc%2FApp.tsx

Also you can checkout added test to for more simple example.

## How has this been tested?

Test that was added wasn't initially was throwing error like 'unable to
read xxx.length of undefined` from somewhere of compressed
`text-mask-core`.

After change it passed well. DatePicker example from above also works
fine after the fix.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.